thesis_code/pyproject.toml
[tool.poetry]
name = "thesis_code"
version = "0.1.0"
description = "dev"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.9.dev0"
#dbt-snowflake = "1.4.*"
snowflake-snowpark-python = ">=1.0.0"
python-dotenv = "^1.0.0"
pandas = "2.0.0"
snowflake-connector-python = {extras = ["pandas"], version = "^3.2.1"}
pyspark = "^3.5.0"
jsonlines = "^4.0.0"
scikit-learn = "^1.3.1"
openai = "^0.28.1"
tokenizers = "^0.15.0"
regex = "2023.10.3"
safetensors = "0.4.0"
sentencepiece = "0.1.99"
transformers = "4.35.2"
MarkupSafe = "2.1.3"
jinja2 = "3.1.2"
mpmath = "1.3.0"
networkx = "3.1"
pillow = "10.1.0"
sympy = "1.12"
torch = "2.1.0"
torchvision = "0.16.0"
pytorch-lightning = "^2.1.1"
nltk = "^3.8.1"
huggingface-hub = "^0.19.3"
datasets = "^2.15.0"
accelerate = "^0.20.1"
psutil = "^5.9.6"
rouge-score = "^0.1.2"
tensorboardx = "^2.6.2.2"
deepl = "^1.16.1"
googletrans = "4.0.0-rc1"
openpyxl = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
